Renamed parseTest.py to parseHPMC.py
